chore: 🤖 add barrel pattern and eslint rule for ordering import
