+++ .eslintrc
@@ -21,14 +21,15 @@
   "settings": {
     "import/resolver": {
       "typescript": {
-        "project": "/*/tsconfig.json"
+        "project": "/*/tsconfig.json",
+        "alwaysTryTypes": true
       },
       "node": {
         "extensions": [".js", ".jsx", ".json", ".ts", ".tsx"]
       }
     }
   },
-  "plugins": ["react", "import", "unused-imports"],
+  "plugins": ["react", "import", "unused-imports", "import"],
   "rules": {
     "react/jsx-uses-react": "off",
     "react/react-in-jsx-scope": "off",
@@ -38,6 +39,28 @@
       {
         "component": true,
         "html": true
+      }
+    ],
+    "import/order": [
+      "error",
+      {
+        "groups": ["builtin", "external", "parent", "sibling", "index", "object", "type"],
+        "pathGroups": [
+          {
+            "pattern": "{react,react-dom/**,react-router-dom}",
+            "group": "builtin",
+            "position": "before"
+          },
+          {
+            "pattern": "@alias/**",
+            "group": "parent",
+            "position": "before"
+          }
+        ],
+        "alphabetize": {
+          "order": "asc"
+        },
+        "newlines-between": "always"
       }
     ]
   }
