+++ package.json
@@ -11,7 +11,7 @@
     "fix:eslint": "eslint --ext .ts,.tsx --fix src",
     "lint:stylelint": "stylelint **/*.{css,scss,sass}",
     "fix:stylelint": "stylelint **/*.{css,scss,sass} --fix",
-    "test": "vitest",
+    "test": "vitest --ui",
     "coverage": "vitest run --coverage",
     "storybook": "start-storybook -p 6006",
     "build-storybook": "build-storybook",
@@ -57,12 +57,13 @@
     "@typescript-eslint/parser": "^5.47.1",
     "@vitejs/plugin-react": "^3.0.0",
     "@vitejs/plugin-react-swc": "^3.0.0",
+    "@vitest/ui": "^0.28.4",
     "babel-loader": "^8.3.0",
     "eslint": "^8.0.1",
     "eslint-config-prettier": "^8.5.0",
     "eslint-config-standard-with-typescript": "^24.0.0",
-    "eslint-import-resolver-typescript": "^3.5.2",
-    "eslint-plugin-import": "^2.26.0",
+    "eslint-import-resolver-typescript": "^3.5.3",
+    "eslint-plugin-import": "^2.27.5",
     "eslint-plugin-jsx-a11y": "^6.6.1",
     "eslint-plugin-n": "^15.0.0",
     "eslint-plugin-promise": "^6.0.0",
