+++ src/App.tsx
@@ -1,6 +1,8 @@
 import { FC } from 'react'
-import { AppRoutes } from '@/routes'
+
 import { AppProvider } from './providers'
+
+import { AppRoutes } from '@/routes'
 
 export const App: FC = () => {
   return (
