+++ src/components/functional/Form/FieldWrapper/index.tsx
@@ -1,7 +1,8 @@
 /** @jsxImportSource @emotion/react */
+import { css } from '@emotion/react'
 import { FC, ReactNode } from 'react'
 import { FieldError } from 'react-hook-form'
-import { css } from '@emotion/react'
+
 import { colors } from '@/constants/colors'
 
 type FieldWrapperProps = {
