+++ src/components/functional/Form/index.ts
@@ -1,3 +1,3 @@
-export { Form } from '@/components/functional/Form/Form'
-export { FieldWrapper } from '@/components/functional/Form/FieldWrapper'
-export { InputField } from '@/components/functional/Form/InputField'
+export * from './Form'
+export * from './FieldWrapper'
+export * from './InputField'
