+++ src/hooks/useAuth.ts
@@ -1,7 +1,8 @@
-import { axios } from '@/libs/axios'
-import { LoginFormValues } from '@/pages/Login/LoginPage'
 import { useNavigate } from 'react-router-dom'
-import { SignupFormValues } from "@/pages/Signup/SignupPage";
+
+import { axios } from '@/libs/axios'
+import { LoginFormValues } from '@/pages/Login/type'
+import { SignupFormValues } from '@/pages/Signup/type'
 
 export const useAuth = () => {
   const navigate = useNavigate()
@@ -10,12 +11,12 @@ export const useAuth = () => {
     await axios.post('/v1/user/login', values).then(() => navigate('/'))
   }
 
-  const signup = async (values: SignupFormValues): Promise<void>  => {
+  const signup = async (values: SignupFormValues): Promise<void> => {
     await axios.post('/v1/user', values).then(() => navigate('/'))
   }
 
   return {
     login,
-    signup
+    signup,
   }
 }
