+++ src/main.tsx
@@ -1,5 +1,6 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
+
 import { App } from '@/App'
 
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
