+++ src/pages/Login/Login.stories.tsx
@@ -1,7 +1,9 @@
-import type { ComponentMeta, ComponentStoryObj } from '@storybook/react'
-import { LoginPage } from '@/pages/Login'
 import { userEvent, within } from '@storybook/testing-library'
 import { MemoryRouter } from 'react-router-dom'
+
+import type { ComponentMeta, ComponentStoryObj } from '@storybook/react'
+
+import { LoginPage } from '@/pages/Login'
 
 export default { component: LoginPage } as ComponentMeta<typeof LoginPage>
