+++ src/pages/Login/Login.test.tsx
@@ -1,0 +1,10 @@
+import { describe } from 'vitest'
+
+describe('something truthy and falsy', () => {
+  test('true to be true', () => {
+    expect(true).toBe(true)
+  })
+  test('test to be true', () => {
+    expect(true).toBe(true)
+  })
+})
