--- src/pages/Login/Logint.test.tsx
@@ -1,7 +1,0 @@
-import { describe, it, expect } from 'vitest'
-
-describe('something truthy and falsy', () => {
-  it('true to be true', () => {
-    expect(true).toBe(true)
-  })
-})
