+++ src/pages/Login/index.ts
@@ -1,0 +1,2 @@
+export * from './Login'
+export * from './type'
