+++ src/pages/NotFound/NotFound.stories.tsx
@@ -1,6 +1,8 @@
-import type { ComponentMeta, ComponentStoryObj } from '@storybook/react'
-import { NotFoundPage } from '@/pages/NotFound'
 import { MemoryRouter } from 'react-router-dom'
+
+import type { ComponentMeta, ComponentStoryObj } from '@storybook/react'
+
+import { NotFoundPage } from '@/pages/NotFound'
 
 export default { component: NotFoundPage } as ComponentMeta<typeof NotFoundPage>
