+++ src/pages/Signup/index.ts
@@ -1,0 +1,2 @@
+export * from './Signup'
+export * from './type'
