+++ src/pages/index.ts
@@ -1,0 +1,4 @@
+export * from './Home'
+export * from './Login'
+export * from './NotFound'
+export * from './Signup'
